refactor: ゲーム状態モジュールを分割
