+++ src/game/state/index.ts
@@ -1,0 +1,6 @@
+export { createEnemies } from './enemy';
+export { prepMaze, initState } from './core';
+export type { GameState, State } from './core';
+export { createFirstStage, nextStageState, restartRun } from './stage';
+export { reducer } from './reducer';
+export type { Action } from './reducer';
